Update container cardinality handling and remove deprecated methods
- Changed the Size field in the container struct from uint16 to uint32 for improved capacity handling.
- Removed deprecated methods for determining conversion thresholds between array and bitmap types.
- Streamlined conversion logic to utilize direct size comparisons, enhancing clarity and maintainability.
- Updated related methods to ensure consistent cardinality updates during type transitions.

+++ container_run.go
@@ -239,7 +239,7 @@ func (c *container) runToArray() {
 	// Create new array data
 	c.Data = make([]byte, len(values)*2)
 	c.Type = typeArray
-	c.Size = uint16(len(values)) // Set cardinality
+	c.Size = uint32(len(values)) // Set cardinality
 	array := c.arr()
 	copy(array, values)
 }
@@ -264,5 +264,5 @@ func (c *container) runToBmp() {
 	}
 
 	// Update cardinality from bitmap
-	c.Size = uint16(bm.Count())
+	c.Size = uint32(bm.Count())
 }
